[#1821] Add config properties for TLS and/or client auth
The Prometheus based resource limits checker can now be configured to
use TLS and provide credentials when connecting to Prometheus' query
API.

Fixes #1821
