+++ core/src/main/java/org/eclipse/hono/config/ClientConfigProperties.java
@@ -13,18 +13,12 @@
 
 package org.eclipse.hono.config;
 
-import java.io.FileInputStream;
-import java.io.IOException;
-import java.util.Objects;
-import java.util.Optional;
-import java.util.Properties;
-
 import org.eclipse.hono.util.Constants;
 
 /**
  * Common configuration properties required for accessing an AMQP 1.0 container.
  */
-public class ClientConfigProperties extends AbstractConfig {
+public class ClientConfigProperties extends AuthenticatingClientConfigProperties {
 
     /**
      * The default amount of time (milliseconds) to wait for an AMQP connection to
@@ -76,31 +70,24 @@ public class ClientConfigProperties extends AbstractConfig {
 
     private String amqpHostname;
     private int connectTimeoutMillis = DEFAULT_CONNECT_TIMEOUT;
-    private String credentialsPath;
     private long flowLatency = DEFAULT_FLOW_LATENCY;
-    private String host = "localhost";
-    private boolean hostnameVerificationRequired = true;
     private int idleTimeoutMillis = DEFAULT_IDLE_TIMEOUT;
     private int initialCredits = DEFAULT_INITIAL_CREDITS;
     private long linkEstablishmentTimeout = DEFAULT_LINK_ESTABLISHMENT_TIMEOUT;
     private String name;
-    private char[] password;
-    private int port = Constants.PORT_AMQPS;
     private int reconnectAttempts = -1;
     private long reconnectMinDelayMillis = DEFAULT_RECONNECT_MIN_DELAY;
     private long reconnectMaxDelayMillis = DEFAULT_RECONNECT_MAX_DELAY;
     private long reconnectDelayIncrementMillis = DEFAULT_RECONNECT_DELAY_INCREMENT;
     private long requestTimeoutMillis = DEFAULT_REQUEST_TIMEOUT;
     private long sendMessageTimeoutMillis = DEFAULT_SEND_MESSAGE_TIMEOUT;
-    private boolean tlsEnabled = false;
-    private String username;
-    private String serverRole = "unknown";
 
     /**
      * Creates new properties with default values.
      */
     public ClientConfigProperties() {
         super();
+        setPort(Constants.PORT_AMQPS);
     }
 
     /**
@@ -112,187 +99,19 @@ public class ClientConfigProperties extends AbstractConfig {
         super(otherProperties);
         this.amqpHostname = otherProperties.amqpHostname;
         this.connectTimeoutMillis = otherProperties.connectTimeoutMillis;
-        this.credentialsPath = otherProperties.credentialsPath;
         this.flowLatency = otherProperties.flowLatency;
-        this.host = otherProperties.host;
-        this.hostnameVerificationRequired = otherProperties.hostnameVerificationRequired;
         this.idleTimeoutMillis = otherProperties.idleTimeoutMillis;
         this.initialCredits = otherProperties.initialCredits;
         this.linkEstablishmentTimeout = otherProperties.linkEstablishmentTimeout;
         this.name = otherProperties.name;
-        this.password = otherProperties.password;
-        this.port = otherProperties.port;
         this.reconnectAttempts = otherProperties.reconnectAttempts;
         this.reconnectMinDelayMillis = otherProperties.reconnectMinDelayMillis;
         this.reconnectMaxDelayMillis = otherProperties.reconnectMaxDelayMillis;
         this.reconnectDelayIncrementMillis = otherProperties.reconnectDelayIncrementMillis;
         this.requestTimeoutMillis = otherProperties.requestTimeoutMillis;
         this.sendMessageTimeoutMillis = otherProperties.sendMessageTimeoutMillis;
-        this.serverRole = otherProperties.serverRole;
-        this.tlsEnabled = otherProperties.tlsEnabled;
-        this.username = otherProperties.username;
     }
 
-    /**
-     * Gets the name or literal IP address of the host that the client is configured to connect to.
-     * <p>
-     * The default value of this property is "localhost".
-     *
-     * @return The host name.
-     */
-    public final String getHost() {
-        return host;
-    }
-
-    /**
-     * Sets the name or literal IP address of the host that the client should connect to.
-     * 
-     * @param host The host name or IP address.
-     * @throws NullPointerException if host is {@code null}.
-     */
-    public final void setHost(final String host) {
-        this.host = Objects.requireNonNull(host);
-    }
-
-    /**
-     * Gets the TCP port of the server that this client is configured to connect to.
-     * <p>
-     * The default value of this property is {@link Constants#PORT_AMQPS}.
-     * 
-     * @return The port number.
-     */
-    public final int getPort() {
-        return port;
-    }
-
-    /**
-     * Sets the TCP port of the server that this client should connect to.
-     * <p>
-     * The default value of this property is {@link Constants#PORT_AMQPS}.
-     * 
-     * @param port The port number.
-     * @throws IllegalArgumentException if port &lt; 1000 or port &gt; 65535.
-     */
-    public final void setPort(final int port) {
-        if (isValidPort(port)) {
-            this.port = port;
-        } else {
-            throw new IllegalArgumentException("invalid port number");
-        }
-    }
-
-    /**
-     * Gets the user name that is used when authenticating to the Hono server.
-     * <p>
-     * This method returns the value set using the <em>setUsername</em> method
-     * if the value is not {@code null}. Otherwise, the user name is read from the
-     * properties file indicated by the <em>credentialsPath</em> property (if not
-     * {@code null}).
-     * 
-     * @return The user name or {@code null} if not set.
-     */
-    public final String getUsername() {
-        if (username == null) {
-            loadCredentials();
-        }
-        return username;
-    }
-
-    /**
-     * Sets the user name to use when authenticating to the Hono server.
-     * <p>
-     * If not set then this client will not try to authenticate to the server.
-     * 
-     * @param username The user name.
-     */
-    public final void setUsername(final String username) {
-        this.username = username;
-    }
-
-    /**
-     * Gets the password that is used when authenticating to the Hono server.
-     * <p>
-     * This method returns the value set using the <em>setPassword</em> method
-     * if the value is not {@code null}. Otherwise, the password is read from the
-     * properties file indicated by the <em>credentialsPath</em> property (if not
-     * {@code null}).
-     * 
-     * @return The password or {@code null} if not set.
-     */
-    public final String getPassword() {
-        if (password == null) {
-            loadCredentials();
-        }
-        return password == null ? null : String.valueOf(password);
-    }
-
-    /**
-     * Sets the password to use in conjunction with the user name when authenticating
-     * to the Hono server.
-     * <p>
-     * If not set then this client will not try to authenticate to the server.
-     * 
-     * @param password The password.
-     */
-    public final void setPassword(final String password) {
-        if (password != null) {
-            this.password = password.toCharArray();
-        } else {
-            this.password = null;
-        }
-    }
-
-    /**
-     * Gets the file system path to a properties file containing the
-     * credentials for authenticating to the server.
-     * <p>
-     * The file is expected to contain a <em>username</em> and a
-     * <em>password</em> property, e.g.
-     * <pre>
-     * username=foo
-     * password=bar
-     * </pre>
-     * 
-     * @return The path or {@code null} if not set.
-     */
-    public final String getCredentialsPath() {
-        return credentialsPath;
-    }
-
-    /**
-     * Sets the file system path to a properties file containing the
-     * credentials for authenticating to the server.
-     * <p>
-     * The file is expected to contain a <em>username</em> and a
-     * <em>password</em> property, e.g.
-     * <pre>
-     * username=foo
-     * password=bar
-     * </pre>
-     * 
-     * @param path The path to the properties file.
-     */
-    public final void setCredentialsPath(final String path) {
-        this.credentialsPath = path;
-    }
-
-    private void loadCredentials() {
-
-        if (username == null && password == null && credentialsPath != null) {
-            try (FileInputStream fis = new FileInputStream(credentialsPath)) {
-                LOG.info("loading credentials for [{}:{}, role: {}] from [{}]",
-                        host, port, serverRole, credentialsPath);
-                final Properties props = new Properties();
-                props.load(fis);
-                this.username = props.getProperty("username");
-                this.password = Optional.ofNullable(props.getProperty("password"))
-                        .map(pwd -> pwd.toCharArray()).orElse(null);
-            } catch (IOException e) {
-                LOG.warn("could not load client credentials for [{}:{}, role: {}] from file [{}]",
-                        host, port, serverRole, credentialsPath, e);
-            }
-        }
-    }
     /**
      * Gets the name being indicated as the <em>container-id</em> in the client's AMQP <em>Open</em> frame.
      * 
@@ -499,67 +318,6 @@ public class ClientConfigProperties extends AbstractConfig {
     }
 
     /**
-     * Checks if the <em>host</em> property must match the distinguished or
-     * any of the alternative names asserted by the server's certificate when
-     * connecting using TLS.
-     * 
-     * @return {@code true} if the host name will be matched against the
-     *         asserted names.
-     */
-    public final boolean isHostnameVerificationRequired() {
-        return hostnameVerificationRequired;
-    }
-
-    /**
-     * Sets whether the <em>host</em> property must match the distinguished or
-     * any of the alternative names asserted by the server's certificate when
-     * connecting using TLS.
-     * <p>
-     * Verification is enabled by default, i.e. the connection will be established
-     * only if the server presents a certificate that has been signed by one of the
-     * client's trusted CAs and one of the asserted names matches the host name that
-     * the client used to connect to the server.
-     * 
-     * @param hostnameVerificationRequired {@code true} if the host name should be matched.
-     */
-    public final void setHostnameVerificationRequired(final boolean hostnameVerificationRequired) {
-        this.hostnameVerificationRequired = hostnameVerificationRequired;
-    }
-
-    /**
-     * Checks if the client should use TLS to verify the server's identity
-     * and encrypt the connection.
-     * <p>
-     * Verification is disabled by default. Setting the <em>trustStorePath</em>
-     * property enables verification of the server identity implicitly and the
-     * value of this property is ignored.
-     * 
-     * @return {@code true} if the server identity should be verified.
-     */
-    public final boolean isTlsEnabled() {
-        return tlsEnabled;
-    }
-
-    /**
-     * Sets whether the client should use TLS to verify the server's identity
-     * and encrypt the connection.
-     * <p>
-     * Verification is disabled by default. Setting the <em>trustStorePath</em>
-     * property enables verification of the server identity implicitly and the
-     * value of this property is ignored.
-     * <p>
-     * This property should be set to {@code true} if the server uses a certificate
-     * that has been signed by a CA that is contained in the standard trust store
-     * that the JVM is configured to use. In this case the <em>trustStorePath</em>
-     * does not need to be set.
-     * 
-     * @param enabled {@code true} if the server identity should be verified.
-     */
-    public final void setTlsEnabled(final boolean enabled) {
-        this.tlsEnabled = enabled;
-    }
-
-    /**
      * Gets the number of attempts (in addition to the original connection attempt)
      * that the client should make in order to establish an AMQP connection with
      * the peer before giving up.
@@ -759,28 +517,5 @@ public class ClientConfigProperties extends AbstractConfig {
         } else {
             this.idleTimeoutMillis = idleTimeoutMillis;
         }
-    }
-
-    /**
-     * Sets the name of the role that the server plays from the client's perspective.
-     * <p>
-     * The default value of this property is <em>unknown</em>.
-     * 
-     * @param roleName The name.
-     * @throws NullPointerException if name is {@code null}.
-     */
-    public final void setServerRole(final String roleName) {
-        this.serverRole = Objects.requireNonNull(roleName);
-    }
-
-    /**
-     * Gets the name of the role that the server plays from the client's perspective.
-     * <p>
-     * The default value of this property is <em>unknown</em>.
-     * 
-     * @return The name.
-     */
-    public final String getServerRole() {
-        return serverRole;
     }
 }
