+++ site/documentation/content/admin-guide/hono-client-configuration.md
@@ -40,9 +40,8 @@ delimiter,  e.g. the variable prefix `HONO_CREDENTIALS` corresponds to the comma
 | `${PREFIX}_KEYSTOREPATH`<br>`--${prefix}.keyStorePath` | no | - | The absolute path to the Java key store containing the private key and certificate that the client should use for authenticating to the server. Either this variable or the `${PREFIX}_KEYPATH` and `${PREFIX}_CERTPATH` variables need to be set in order to enable *SASL External* based authentication to the server. The key store format can be either `JKS` or `PKCS12` indicated by a `.jks` or `.p12` file suffix respectively. |
 | `${PREFIX}_LINKESTABLISMENTTIMEOUT`<br>`--${prefix}.linkEstablishmentTimeout` | no | `1000` | The maximum amount of time (milliseconds) that the client should wait for the service's *attach* frame during link establishment. This property can be used to tune the time period to wait according to the network latency involved with the communication link between the client and the service. |
 | `${PREFIX}_NAME`<br>`--${prefix}.name` | no | - | The name to use as the *container-id* in the client's AMQP *open* frame during connection establishment. |
-| `${PREFIX}_PORT`<br>`--${prefix}.port` | no | `5671` | The port that the service is listening on. |
-| `${PREFIX}_USERNAME`<br>`--${prefix}.username` | no | - | The username to use for authenticating to the service. This property (and the corresponding *password*) needs to be set in order to enable *SASL Plain* based authentication to the service.|
 | `${PREFIX}_PASSWORD`<br>`--${prefix}.password` | no | - | The password to use for authenticating to the service. |
+| `${PREFIX}_PORT`<br>`--${prefix}.port` | no | `5671` | The port that the service is listening on. |
 | `${PREFIX}_SENDMESSAGETIMEOUT`<br>`--${prefix}.sendMessageTimeout` | no | `1000` | The maximum number of milliseconds to wait for a delivery update after an event or command message was sent before the send operation is failed. Setting this value to a higher value increases the chance of successful service invocation in situations where network latency is high. |
 | `${PREFIX}_RECONNECTATTEMPTS`<br>`--${prefix}.reconnectAttempts` | no | `-1` | The number of attempts (in addition to the original connection attempt) that the client should make in order to establish an AMQP connection with the peer before giving up. The default value of this property is -1 which means that the client will try forever. |
 | `${PREFIX}_RECONNECTDELAYINCREMENT`<br>`--${prefix}.reconnectDelayIncrement` | no | `100` | The factor (milliseconds) used in the exponential backoff algorithm for determining the delay before trying to re-establish an AMQP connection with the peer. The delay after an initial, failed connection attempt will be the value of the `${PREFIX}_RECONNECTMINDELAY` variable. Each subsequent connection attempt will use a random delay between the minimum delay and the value determined by exponentially increasing the delay by the `${PREFIX}_RECONNECTDELAYINCREMENT` factor. The overall limit of the delay time is defined by the `${PREFIX}_RECONNECTMAXDELAY` variable. |
@@ -51,9 +50,9 @@ delimiter,  e.g. the variable prefix `HONO_CREDENTIALS` corresponds to the comma
 | `${PREFIX}_REQUESTTIMEOUT`<br>`--${prefix}.requestTimeout` | no | `200` | The maximum number of milliseconds to wait for a response before a service invocation is failed. Setting this value to a higher value increases the chance of successful service invocation in situations where network latency is high. |
 | `${PREFIX}_SECUREPROTOCOLS`<br>`--${prefix}.secureProtocols` | no | `TLSv1.2` | A (comma separated) list of secure protocols that are supported when negotiating TLS sessions. Please refer to the [vert.x documentation](https://vertx.io/docs/vertx-core/java/#ssl) for a list of supported protocol names. |
 | `${PREFIX}_TLSENABLED`<br>`--${prefix}.tlsEnabled` | no | `false` | If set to `true` the connection to the peer will be encrypted using TLS and the peer's identity will be verified using the JVM's configured standard trust store.<br>This variable only needs to be set to enable TLS explicitly if no specific trust store is configured using the `${PREFIX}_TRUSTSTOREPATH` variable. |
-| `${PREFIX}_SNISERVERNAME`<br>`--${prefix}.SniServerName` | no | - | Explicitly override the host to use for the TLS SNI server name. If SNI server name is not specified, the host from `${PREFIX}_HOST` will be used with SNI. |
 | `${PREFIX}_TRUSTSTOREPATH`<br>`--${prefix}.trustStorePath` | no  | - | The absolute path to the Java key store containing the CA certificates the adapter uses for authenticating the service. This property **must** be set if the service has been configured to support TLS. The key store format can be either `JKS`, `PKCS12` or `PEM` indicated by a `.jks`, `.p12` or `.pem` file suffix respectively. |
 | `${PREFIX}_TRUSTSTOREPASSWORD`<br>`--${prefix}.trustStorePassword` | no | - | The password required to read the contents of the trust store. |
+| `${PREFIX}_USERNAME`<br>`--${prefix}.username` | no | - | The username to use for authenticating to the service. This property (and the corresponding *password*) needs to be set in order to enable *SASL Plain* based authentication to the service.|
 
 ## Response Caching
